refactored tracks library
uses spotify solely for spotify api get
