+++ server/models/track.py
@@ -16,6 +16,7 @@ import requests
 
 from server import settings
 from server.lib import tokens
+from server.lib import spotify
 
 connect(settings.DB)
 
@@ -73,6 +74,7 @@ class Track(Document):
     track_number = IntField()
     type = StringField()
     uri = StringField()
+    image = DictField()
 
     def to_log(self):
         dict = {
@@ -121,3 +123,10 @@ class Track(Document):
         if tracks:
             return True
         return False
+
+    # TODO: Call track/{id} to get image url
+    def set_image(self):
+        track_url = 'tracks/{id}'.format(id=self.track_id)
+        track_data = spotify.get(settings.API_URL_BASE.format(endpoint=track_url))
+        import pdb
+        pdb.set_trace()
